+++ RentAll.Infrastructure/Repositories/Contacts/Update.cs
@@ -15,7 +15,7 @@ namespace RentAll.Infrastructure.Repositories.Contacts
             {
                 ContactId = contact.ContactId,
                 ContactCode = contact.ContactCode,
-                ContactTypeId = contact.ContactTypeId,
+                ContactTypeId = (int)contact.ContactTypeId,
                 FirstName = contact.FirstName,
                 LastName = contact.LastName,
                 Address1 = contact.Address1,
